requirements.txt
Babel==2.9.0
backports-abc==0.5
backports.functools-lru-cache==1.5
bcrypt==3.1.7
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
contextlib2==0.6.0.post1
cryptography==3.3.1
debtcollector==1.22.0
enum34==1.1.10
funcsigs==1.0.2
futures==3.3.0
idna==2.9
importlib-resources==3.3.1
ipaddress==1.0.23
iso8601==0.1.13
monotonic==1.5
msgpack==1.0.2
netaddr==0.8.0
netifaces==0.10.9
oslo.config==6.2.0
oslo.context==2.23.1
oslo.i18n==3.25.1
oslo.log==3.38.0
oslo.serialization==2.29.2
oslo.utils==3.42.1
paramiko==2.4.2
pathlib2==2.3.5
pbr==5.5.1
psycopg2==2.7.5
pyasn1==0.4.8
pycparser==2.20
pycrypto==2.6.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.23.0
rfc3986==1.4.0
scandir==1.10.0
singledispatch==3.4.0.3
six==1.14.0
soupsieve==1.9.4
stevedore==1.32.0
tornado==4.4.2
typing==3.7.4.3
urllib3==1.25.9
wrapt==1.12.1
zipp==1.2.0
